Improve manifest handling for live builds.
We can respect the manifest and whether files are dirty for the initial
build before starting the --live server which is essential to ensure
that {{livereload}} parameters are replaced with the new URL for the
websocket server endpoint but it is not necessary to copy files over if
they are not dirty, we only need to parse pages.

This commit adds a pages_only flag to achieve this goal; when pages_only
is set then it overrides the normal behavior and forces pages to be
re-rendered.