Match URLs for generator destinations.
So that the `link` helper does not error when linking to generator
destination files. This is convoluted as there is no direct mapping
between source files and generators can create additional files such as
copying the underlying JSON data files and creating JSON index files.

It covers all the current generator output destinations but may need
updating as generators change.