Support SSH for init command.
Can now clone via SSH too which is useful for blueprints that should be
private. To init via SSH the --private-key option must be given:

cargo run -- init build/foo/bar git@github.com:hypertext-live/blueprint-newcss.git --private-key=.ssh/id_rsa

Added the rpassword crate so that we can prompt for a passphrase for the
private key when necessary.

We should support a global `~/.hypertext/ssh.toml` file so that
people can configure SSH private keys per domain name to make the
experience better. Also integration with SSH agent would be a nice to
have but not essential right now as use of SSH for cloning blueprints is
not so important.

However when we want to "sync" a project to a remote repository and we
want to support private repositories better SSH config will be more of a
priority.