Improve paths handling.
So that we can now pass arbitrary paths to compile specific sections of
a site and they may also be specified in the `paths` section of a
profile.

Paths must be within the site source configuration; when specified on
the command line then the path is not modified however when paths come
from a profile declaration they are automatically made relative to the
site source so that paths do not have lots of redundant prefixes. For
example:

```toml
[profile.docs]
paths = ["assets", "docs"]
```

Will compile "site/assets" and "site/docs".