Refactor BuildConfig -> ProfileSettings.
The RuntimeOptions now has a ProfileSettings (settings) which inherits
from the root profile settings (currently `[build]`) and then overrides
using a particular profile (when `--profile` is given) before finally
applying the command line arguments.

Requires work to move `extension` into the profile settings so that they
can be changed on a profile by profile basis.

Also fix a small bug with multiple locales when live reload is active -
only the last locale was being compiled, disabled the conditional for
dry_run when multiple locales are active.