Cross-compilation for macos.
Finally get it working, see `build_osx.sh`. Had to switch to
`regex-fancy` from `regex-onig` in the syntect dependency which will
slow down the syntax highlighting library.

Later it would be good to restore onig if we can get the clang linker to
find the onig symbols.

An important change was the switch to using `vendored-openssl` feature
for `git2` so that openssl is built from source for the target platform.

Requires that `osxcross` has been built and target/bin is available in PATH.