Refactor to support multiple search indices.
The bundle, js and wasm settings apply to all search configurations
whilst other settings have been migrated to a hash map so that search
indices are now named. The helper now requires that an `id` hash
parameter is passed to determine the search index to use:

{{search id="posts" script=true}}

Requires work to inherit the id from the hashmap key.