Support generating sitemap.xml files.
All files are placed in the sitemap folder and automatically chunked
based on the number of sitemap `entries` which generates `sitemap/1.xml`
etc. which are referenced by the `sitemap/index.xml` file.

Note that the `lastmod` field for each entry comes from the *source file
modified time* not from the modified time of the file written to disc.

To configure the sitemap:

```toml
[build.sitemap]
entries = 10000
name = "sitemap-folder-name"
```

Typically no configuration is needed and the `name` should only be
changed if there is a conflict with the default `sitemap` folder name.

So to enable sitemaps just create the default configuration:

```toml
[build.sitemap]
```

The default number of entries is 25,000 which is half the maximum documented
amount of 50,000 so we should clear the 50MB hard limit easily assuming
the URLs are not extremely long.

For multi-lingual sites a `sitemap` folder is created in each language
folder and the robots.txt file lists multiple sitemaps.