Draft TLS support.
Adds the `tls` flag to the web server options which when present asks
the web server to look for environment variables that determine a
certificate and key path to use for the TLS server.

When the flag is set and both the environment variables are declared
then the web server attempts to run in TLS mode. It passes a
flag back to the caller on the bind channel so it can use the https:
scheme instead of the default http: when opening the browser window.

Requires a little work to allow configuring the environment variable
names but appears to work as expected.

Consider redirecting http: requests to https: when the TLS server is
running.

To generate certificates for local testing users are expected to use the
mkcert tool.