Support noindex partial and flag.
Adds a noindex field to the page data which is managed via the standard
`noindex` partial which checks the flag and embeds a meta directive when
present.

Note that the `noindex` partial is included by the standard `head`
partial. If a layout does not include the `head` partial it is their
responsibility to call the `noindex` partial if they want the
functionality.