Support [css.main] setting.
So that it is consistent with the way we manage global scripts.

Configure the settings like this:

[css]
main = ["/assets/css/style.css", "/assets/css/alt.css"]

And they are automatically included for every page when using the
`{{styles}}` helper.

This means that the `styles` helper no longer has a `main` hash
parameter.