Use local version of warp.
So that we can access the `hosts` filter now for multi-host support.
Later, when a new release is available including the new filters we can
switch over.

Note that there were compiler errors which needed to be resolved by
removing `Cargo.lock` so the `hyper` transitive dependency was installed
as a newer version (0.13.7 rather than 0.13.5).