Draft virtual host live reload.
Requires using a TLD other than the vanilla `localhost` as SSL
certificates are not valid using the wildcard `*.localhost` so settled
on the default of `*.localhost.dev` for now.

For projects with workspace members live reload appears to be working as
expected. A helper to link to the local subdomain during development
would be useful.