Draft AOT menu compilation.
Menus are parsed before pages and the compiled versions are re-assigned
to the page data so they can be referenced in each page.

Requires work to:

* Create a menu helper to render the compiled menus.
* Support menus of page references (URL paths)
* Support glob patterns in menu definitions
* Support arbitrary nested menus
* Support referencing all children of current directory

The current version works with manually declared menus (such as MENU.md)
which is the most versatile as it can be arbitrary markup but in many
situations it may be preferable to just refer to a bunch of pages.

Currently to test menu rendering a page can do:

{{{menu.sidebar.result}}}

Note the triple handlebars as menus are always pre-compiled to HTML.