Switch registry to bucket.
The cloudfront caching was causing digest checksum mismatch as it was
returning a cached version of the plugin.

This is only an issue during this development phase when we are force
overwriting plugins with the same version number; later once we always
increment the plugin versions this won't be an issue and we can revert
to cached cloudfront packages.