Refactor document rendering.
So that we just call a top-level `{{document}}` helper which locates
the layout assigned to the page data and renders it. Layouts now need to
call a `{{block}}` wherever they want to embed the current page.

The `partial` helper has been renamed to `render` and expects a `href`
argument:

{{render "/docs/"}}

For the case when we want to merge multiple documents into a single page
for print media.