Support reloading front matter in live reload.
So that changing page data is reflected in the browser when live reload
is enabled. This can be verified by changing the `title` in the front
matter and seeing it update on a page.

See the notes in `renderer.rs` for why we must use `append()` on the
page data and cannot directly assign:

*page_write = computed_page;

Which would overwrite the internal, private computed page data.

Closes #189.