Support following edits in the browser.
When `follow-edits` is enabled:

```toml
[livereload]
follow-edits = true
```

Then we send the page href to the browser so that it navigates to the
last edited page.

This is not yet enabled by default as it is still considered
experimental but later when it is deemed to be stable we should consider
making this enabled by default.