Fix handling of collection index filters.
Remove the logic for filtering by field and instead use glob patterns on
the index definition in an `exclude` list to prevent certain files
appearing in the index.

Closes #307. This bug was introduced when changing the behavior of the
`listing` field for directory listings and this field should only be
used for directory listings and not to omit pages from collections
indices. This fix removes the conflation.

Comment out unused functions in the json_path utilities.