Use rayon::scope() for render channel.
This gives us the link between the web server requests and dynamically
rendering when pages are requested during live reload.

The issue with using `Arc<Mutex<Invalidator>>` is that `MutexGuard` is
not `Send` and we get a compiler error as the entire future in the
tokio::task is considered not to be `Send`.

We need to call `await` on the `Updater` which is inside the
`Invalidator` so the only possible solution to this approach would be
replacing `std::sync::Mutex` with `tokio::sync::Mutex` but this cascades
down to all the `RwLock` structs too which would then be problematic
because these require a call to `await` and we access the values behind
locks in a lot of synchronous code.

By using `rayon::scope()` we can access `&mut self` in `render_channel()`
in the `Invalidator` which gives us access to the `Updater` and
performing a re-render.

More information is here (see the error message in the article):

https://tokio.rs/tokio/tutorial/shared-state