Fix web server channel bug.
Trying to launch a web server without live reload enabled was attempting
to configure using the channels and generating an error accessing a
channel that did not exist.

Added the virtual_hosts!() macro so we can branch on virtual host
filters depending upon whether a `watch` flag has been set on the
default virtual host.

Currently, due to the warp types it is not possible to mix virtual
hosts that are `watch` and those that are not.