Draft logic for compiler pinning.
To pin to a specific compiler version users can put a `pin` version in
`site.toml`. The specified version should previously have been
installed using `uvm install`.

If the current executable matches the specified `pin` version no action
is taken.

If the pinned version differs from the current executable we try to
locate an installed version and `fork()` a new process using the pinned
version.

If the pinned version does not match and no executable can be found
matching the pinned version an error message is shown indicating that
the requested version is not installed and the user should use `uvm
install`.