Isolate plugins by release version.
They are installed into the directory for each release, for example:

~/.uwe/releases/0.42.8/plugins

Need to use create_dir_all() when testing using `cargo run` and the
installation directory for the executable version does not exist yet.

The `current_release_dir()` and `plugins_dir()` functions should only be
called after `opts::project_path()` so that we can be certain the
executable version used to determine the release directory is correct.

Closes #389.