Support `upm remove` command.
To delete an installed plugin. Accepts a plugin spec and finds all
plugins that match in the registry; for each plugin that matches and is
installed delete it.

Also migrate ReleaseVersion -> VersionKey so that it can be used in
other code that manages a list of versions (eg: the plugin registry).