Bug fix for link catalog path.
And fixes for an issue that caused plugins with scoped plugins to be
stale and cause some projects to break.

The fix was to re-publish updated registry items without `null` values as
we are no longer using `Option<Vec<_>>` and preferring `Vec<_>` and then
skipping serialization when the vector is empty.