Add documentation to release process.
Compiles the `docs` profile for the website, copies it to the
`documetnation` repository and pushes the changes.

Then the `plugin.toml` is updated with the new version and the plugin is
published.

Now we just need work on the install end to resolve `std::documentation`
using the release version otherwise fallback to the latest version.