Draft refactor to add blueprint to deps.
So that when a new project is created the generated site.toml file
includes the blueprint as a dependency.

Currently, only tested with projects created from registry plugins,
and local folder paths requires work to test `git` URLs - which needs
support for an optional `--git-path` which specifies a sub-folder within
a git repository to load the plugin from.