Assign page query data in live reload.
This is the first step to improving the live reload and collections
integration. At least now the data that has been assigned should not
disappear whilst using live reload.

However, there is still work to be done to invalidate collections
indices for pages (and data documents) so that the data is not stale.

Also, handling the `each` operation will be tricky as it will be hard to
know which stale pages should be removed if we re-generate pages using
the `each` operation.