Draft logic that rebuilds collections database.
So that when documents in the `collections` folder change then the
entire collections database is re-loaded and the indices are re-built
which means that collections using JSON/TOML files will work as expected
with live reload.

Requires more effort to ensure that collections providers using front
matter (`pages`) can be invalidated.