Improve logic for plugin styles/scripts collation.
So that a dependency does not request all `styles` or `scripts`. The
filters are respected when adding files to the collation.

This means that a dependency like this:

```toml
[dependencies."std::syntax"]
version = "^1"
apply = { styles = [ { to = "**", filter = "solarized-light.css" } ] }
```

No longer creates symbolic links for every file in the plugin `styles`
folder but only for those that have matches.

If any apply pattern matches everything without a file (`**`) then they
will all be collated.