Fix dev server race condition.
When starting the development server with live reload enabled each
virtual host starts a thread that watches the file system for changes to
the source directory for that project.

When running with multiple hosts sometimes the server would start before
all of the watch channels were prepared. Because the web server
filters expect that it has a clear request/response channels for SSR if
the watch channels were not prepared the web server would block waiting
for a channel that would never receive (live_render).

By ensuring that the watch channels are ready beforehand the server no
longer hangs when starting multiple virtual hosts via the dev command.