Disable web server signals in editor command.
So that sending TERM with Ctrl+c when the editor command is running does
not attempt a graceful shutdown but immediately closes the process and
corresponding native window.

Without this update Ctrl+c is handled by the actix server and the native
window hangs whilst the server waits for a graceful shutdown to
complete.