Support UWE_WEBSOCKET_URL environment variable.
This allows us to use live reload when building the editor interface.

Usually the live reload websocket will inspect `document.location.host`
to determine where to connect to which is important for ephemeral ports
but when we are running the editor locally for development we need to
force the websocket URL so it connects to another process.