Improve plugin filtering when path filters.
When path filters are given (for example the website `docs` profile)
then the recent changes to the plugins logic were causing plugins to not
be included in the compiled output as they were removed by the filter
logic.

Improved the filter logic to ensure that plugins with a local file
system path and git URL are always included in the list of files to
compile even when path filters are given.

Closes #499.